refactor: ImageDownloader 객체 생성
