+++ Diary/Common/Util/ImageCacheManager.swift
@@ -8,9 +8,6 @@
 import UIKit
 
 final class ImageCacheManager {
-    static let shared = ImageCacheManager()
-    private init() { }
-    
     private let cache = NSCache<NSString, UIImage>()
     
     func set(object: UIImage, forKey key: String) {
